+++ bliss/shell/cli/__init__.py
@@ -7,3 +7,7 @@
 
 from .repl import *
 from .main import *
+
+# boolean shared beetween TypingHelper in bliss/shell/cli/typing_helper.py
+# and NEWstatus_bar in bliss/shell/cli/ptpython_statusbar_patch.py
+typing_helper_active = True